fix：move qnUpload hepler into service
