+++ app/service/topic.js
@@ -1,6 +1,7 @@
 'use strict';
 
 const Service = require('egg').Service;
+const qiniu = require('qiniu');
 
 class TopicService extends Service {
   /*
@@ -178,6 +179,31 @@ class TopicService extends Service {
     topic.author_id = authorId;
 
     return topic.save();
+  }
+
+  qnUpload(readableStream, key) {
+    const { accessKey, secretKey, bucket } = this.config.qn_access;
+
+    const mac = new qiniu.auth.digest.Mac(accessKey, secretKey);
+    const putPolicy = new qiniu.rs.PutPolicy({ scope: bucket });
+    const uploadToken = putPolicy.uploadToken(mac);
+
+    const config = new qiniu.conf.Config();
+    const formUploader = new qiniu.form_up.FormUploader(config);
+    const putExtra = new qiniu.form_up.PutExtra();
+
+    return new Promise(function(resolve, reject) {
+      formUploader.putStream(uploadToken, key, readableStream, putExtra, function(respErr, respBody, respInfo) {
+        if (respErr) {
+          reject(respErr);
+        }
+        if (respInfo.statusCode === 200) {
+          resolve(respBody);
+        } else {
+          reject(new Error('上传失败:statusCode !== 200'));
+        }
+      });
+    });
   }
 }
